Move literal to expression dir
